[bitcoin] fix integer overflow on hardened derivation (int32_t -> int64_t)

+++ ledger-core/inc/core/key/ExtendedPublicKey.hpp
@@ -48,7 +48,7 @@ namespace ledger {
         template <class NetworkParameters>
         class ExtendedPublicKey {
         public:
-            static inline DeterministicPublicKey _derive(int index, const std::vector<int32_t>& childNums, const DeterministicPublicKey& key) {
+            static inline DeterministicPublicKey _derive(int index, const std::vector<int64_t>& childNums, const DeterministicPublicKey& key) {
                 if (index >= childNums.size()) {
                     return key;
                 }
